backend/requirements/base.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# ASGI server (required for async streaming views)
uvicorn[standard]>=0.34.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# JSON & Data
jsonschema==4.21.1
python-json-logger==2.0.7
sentry-sdk==1.40.0
PyYAML>=6.0.2  # For SKILL.md parsing

# AI Frameworks
pydantic-ai==0.0.18

# Signal Extraction (Phase 1)
sentence-transformers==2.5.1
openai>=1.54.3  # Updated for pydantic-ai compatibility
numpy>=1.26.4  # Allow newer versions for Python 3.13+ compatibility
torch>=2.6.0  # Use latest for Python 3.13 support

# Token-based chunking (Phase 2)
tiktoken>=0.5.2  # Use latest with prebuilt wheels for Python 3.13

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4>=4.12.0  # HTML parsing for URL evidence ingestion

# Vector DB (pgvector — native PostgreSQL, HNSW indexes)
pgvector>=0.2.4

# Graph clustering (Leiden community detection)
python-igraph>=0.11.6
leidenalg>=0.10.2

